refactor: consolidate remote-debugger package into other packages
• Move shared types and utilities to @flyde/core
• Move runtime client functionality to @flyde/runtime
• Move editor client functionality to @flyde/flow-editor
• Move server setup functionality to @flyde/dev-server
• Update all imports across codebase to use new locations
• Remove @flyde/remote-debugger package and dependencies
• Reduce package count from 9 to 8 packages while maintaining functionality

This consolidation improves architecture by placing each component where it's
actually used, avoiding unnecessary dependencies and maintaining clean separation of concerns.
