+++ dev-server/src/debugger/index.ts
@@ -1,0 +1,1 @@
+export * from "./setup-server";
