+++ flow-editor/src/flow-editor/DataInspectionModal/DataInspectionModal.tsx
@@ -4,8 +4,7 @@ import { Input } from "../../ui";
 import { ScrollArea } from "../../ui";
 import { Alert, AlertDescription } from "../../ui";
 import { Button } from "../../ui";
-import { DebuggerEvent, PinDebuggerEvent, PinType } from "@flyde/core";
-import { HistoryPayload } from "@flyde/remote-debugger";
+import { DebuggerEvent, PinDebuggerEvent, PinType, HistoryPayload } from "@flyde/core";
 import React, { useEffect } from "react";
 import { timeAgo, useDebounce } from "../..";
 import { BrowserOnlyReactJson } from "../../lib/analytics-value-renderer/BrowserJsonView";
