+++ pnpm-workspace.yaml
@@ -4,7 +4,6 @@ packages:
   - "dev-server"
   - "stdlib"
   - "website"
-  - "remote-debugger"
   - "resolver"
   - "runtime"
   - "vscode-extension"
