--- remote-debugger/package.json
@@ -1,54 +1,0 @@
-{
-  "name": "@flyde/remote-debugger",
-  "version": "1.0.0-alpha.18",
-  "description": "",
-  "main": "dist/index.js",
-  "module": "dist/esm/index.js",
-  "typings": "dist/index.d.ts",
-  "exports": {
-    ".": {
-      "import": "./dist/esm/index.js",
-      "require": "./dist/index.js",
-      "types": "./dist/index.d.ts"
-    },
-    "./dist/*": {
-      "import": "./dist/esm/*.js",
-      "require": "./dist/*.js",
-      "types": "./dist/*.d.ts"
-    }
-  },
-  "scripts": {
-    "build:ts": "tsc -p .",
-    "build:esm": "tsc -p tsconfig.esm.json",
-    "dev": "pnpm run watch:ts",
-    "bts": "tsc -p .",
-    "build": "rm -rf dist && pnpm run build:ts && pnpm run build:esm",
-    "watch:ts": "tsc -p . -w --incremental"
-  },
-  "files": [
-    "dist/",
-    "README.md"
-  ],
-  "author": "",
-  "license": "MIT",
-  "dependencies": {
-    "@babel/runtime": "^7.22.10",
-    "@flyde/core": "workspace:*",
-    "@socket.io/admin-ui": "^0.2.0",
-    "@types/component-emitter": "^1.2.11",
-    "@types/express": "^4.17.17",
-    "@types/node": "^11.15.54",
-    "@ungap/structured-clone": "^1.2.0",
-    "axios": "^0.27.2",
-    "component-emitter": "^1.3.0",
-    "concurrently": "^4.1.2",
-    "express": "^4.18.2",
-    "flatted": "^3.2.7",
-    "serialize-error": "^8.1.0",
-    "socket.io": "^4.7.2",
-    "socket.io-client": "^4.7.2",
-    "ts-node-dev": "^2.0.0",
-    "typescript": "^5.5.4"
-  },
-  "gitHead": "5f06bd4a0f3da610e64d74c19cc84babc4557f56"
-}
