--- remote-debugger/src/client.ts
@@ -1,2 +1,0 @@
-export * from "./clients/editor";
-export * from "./clients/runtime";
