--- remote-debugger/src/index.ts
@@ -1,3 +1,0 @@
-export * from "./common";
-export * from "./clients/editor";
-export * from "./clients/runtime";
