--- remote-debugger/tsconfig.esm.json
@@ -1,10 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "module": "ES2020",
-    "moduleResolution": "node",
-    "outDir": "dist/esm",
-    "declaration": true,
-    "declarationMap": true
-  }
-}
