--- remote-debugger/tsconfig.json
@@ -1,16 +1,0 @@
-{
-    "extends": "../tsconfig.json",
-    "compilerOptions": {
-		"outDir": "dist",
-		"strict": false,
-		"noUnusedLocals": false,
-		"noUnusedParameters": false
-    },
-    "include": [
-        "src/**/*.ts",
-        "src/**/*.tsx"
-    ],
-    "exclude": [
-		"node_modules"
-	]
-}
