+++ vscode-extension/package.json
@@ -24,8 +24,8 @@
   ],
   "scripts": {
     "vscode:prepackage": "npm run build",
-    "link-all": "npm link @flyde/core @flyde/resolver @flyde/dev-server @flyde/runtime @flyde/remote-debugger @flyde/flow-editor @flyde/stdlib",
-    "update-all": "npm update @flyde/core @flyde/resolver @flyde/dev-server @flyde/runtime @flyde/remote-debugger @flyde/flow-editor",
+    "link-all": "npm link @flyde/core @flyde/resolver @flyde/dev-server @flyde/runtime @flyde/flow-editor @flyde/stdlib",
+    "update-all": "npm update @flyde/core @flyde/resolver @flyde/dev-server @flyde/runtime @flyde/flow-editor",
     "compile": "tsc -p ./",
     "build": "npm run compile && npm run typecheck:webview && npm run build:webview",
     "build:webview": "vite build",
@@ -205,6 +205,7 @@
     "@typescript-eslint/eslint-plugin": "^5.30.0",
     "@typescript-eslint/parser": "^5.30.0",
     "@vitejs/plugin-react": "^4.2.0",
+    "@vscode/sqlite3": "5.1.8-vscode",
     "@vscode/test-electron": "^2.4.1",
     "eslint": "^8.18.0",
     "fs-extra": "^10.1.0",
@@ -219,7 +220,6 @@
     "@flyde/core": "workspace:*",
     "@flyde/dev-server": "workspace:*",
     "@flyde/flow-editor": "workspace:*",
-    "@flyde/remote-debugger": "workspace:*",
     "@flyde/resolver": "workspace:*",
     "@flyde/runtime": "workspace:*",
     "@flyde/stdlib": "workspace:*",
